# [MTV-2318] Improve app startup time in dev mode