--- a/package.json
+++ b/package.json
@@ -89,6 +89,7 @@
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-testing-library": "^5.6.0",
     "eslint-plugin-unused-imports": "^2.0.0",
+    "fork-ts-checker-webpack-plugin": "^9.0.2",
     "husky": "^9.1.7",
     "i18next-parser": "^9.3.0",
     "jest": "^28.1.3",