## 📝 Links
https://issues.redhat.com/browse/MTV-2318

## 📝 Description
Improving the startup time when running MTV locally by using a package that improves type checking speed. This change saves 1 minute and 56 seconds off of the current 2 minutes and 8 seconds it takes to start the app. This new plugin used is known for the following:
- Speeds up [TypeScript](https://github.com/Microsoft/TypeScript) type checking (by moving it to a separate process) 🏎
- Supports modern TypeScript features like [project references](https://www.typescriptlang.org/docs/handbook/project-references.html) and [incremental mode](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-4.html#faster-subsequent-builds-with-the---incremental-flag) ✨
- Displays nice error messages with the [code frame](https://babeljs.io/docs/en/next/babel-code-frame.html) formatter 🌈

So a nice side-effect here is improved readability of error messages too.

## 🎥 Demo
**Before:**
<img width="361" alt="webpack 5 98 0 compiled successfully in 127760 ms" src="https://github.com/user-attachments/assets/b5cbd9ad-7bad-43c3-9519-0b8142370557" />

**After:**
<img width="361" alt="webpack 5 98 0 compiled successfully in 11190 ms" src="https://github.com/user-attachments/assets/92cbda65-9ce1-4409-9278-b691819148ff" />